fix visualization
