+++ deploy/cpp_infer/include/utility.h
@@ -26,11 +26,19 @@
 #include <fstream>
 #include <numeric>
 
+#include "opencv2/core.hpp"
+#include "opencv2/imgcodecs.hpp"
+#include "opencv2/imgproc.hpp"
+
 namespace PaddleOCR {
 
 class Utility {
 public:
   static std::vector<std::string> ReadDict(const std::string &path);
+
+  static void
+  VisualizeBboxes(const cv::Mat &srcimg,
+                  const std::vector<std::vector<std::vector<int>>> &boxes);
 
   template <class ForwardIterator>
   inline static size_t argmax(ForwardIterator first, ForwardIterator last) {
