+++ deploy/cpp_infer/src/main.cpp
@@ -53,7 +53,8 @@ int main(int argc, char **argv) {
   DBDetector det(config.det_model_dir, config.use_gpu, config.gpu_id,
                  config.gpu_mem, config.cpu_math_library_num_threads,
                  config.max_side_len, config.det_db_thresh,
-                 config.det_db_box_thresh, config.det_db_unclip_ratio);
+                 config.det_db_box_thresh, config.det_db_unclip_ratio,
+                 config.visualize);
   CRNNRecognizer rec(config.rec_model_dir, config.use_gpu, config.gpu_id,
                      config.gpu_mem, config.cpu_math_library_num_threads,
                      config.char_list_file);
